Beautiful code

+++ bot.py
@@ -1,9 +1,6 @@
-
 from Class import *
 import os
-
 from discord import app_commands
-
 
 Bot = bot()
 
@@ -16,35 +13,23 @@ async def on_ready():
 
 
 @Bot.command(pass_context=True)
-async def id(ctx):
-    await ctx.channel.send(ctx.message.author.id)
-
-
-@Bot.command(pass_context=True)
-async def prepare(ctx):
+async def prepare(ctx, message):
     if ctx.message.author.id == 343279631807545356:
-        await ctx.channel.send(
-            embed=discord.Embed(colour=0x39d0d6, title="Room management panel",
-                                description="Name - change room name \n\n Lock - lock room \n\n Unlock - unlock room \n\n SCP - Set count people"),
-            view=PersistentView()
-        )
-        await ctx.message.delete()
+        if message=="0":
+            await ctx.channel.send(
+                embed=discord.Embed(colour=0x39d0d6, title="Room management panel",
+                                    description="Name - change room name \n\n Lock - lock room \n\n Unlock - unlock room \n\n SCP - Set count people"),
+                view=PersistentView()
+            )
+            await ctx.message.delete()
+        if message=="1":
+            await ctx.channel.send(
+                embed=discord.Embed(colour=0x39d0d6, title="Тренувальник для переводу чисел в різні системи числення",
+                                    description="Оберіть в яку систему числення ви хочете переводити число"),
+                view=PersistentViewtest()
+            )
+            await ctx.message.delete()
 
-
-def det(n, arr):
-    print(arr)
-    if n == 2:
-        return arr[0][0] * arr[1][1] - arr[0][1] * arr[1][0]
-
-    ans = 0
-    for i in range(n):
-        arrcp = []
-        for i1 in range(1, n):
-            arrcp.append([arr[i1][j] for j in range(n) if (j != i)])
-        ans += arr[0][i] * pow(-1, i + 1 + 1) * det(n - 1, arrcp)
-        print("ANSWER")
-        print(ans)
-    return ans
 
 
 @Bot.command(pass_context=True)
@@ -57,20 +42,8 @@ async def determinant(ctx, size, matrix):
     await ctx.defer()
     await ctx.reply(det(n, arr))
 
-@Bot.command(pass_context=True)
-async def calc(ctx, message):
-    await ctx.message.reply(eval(message))
 
 
-@Bot.command(pass_context=True)
-async def prepare_test(ctx):
-    if ctx.message.author.id == 343279631807545356:
-        await ctx.channel.send(
-            embed=discord.Embed(colour=0x39d0d6, title="Тренувальник для переводу чисел в різні системи числення",
-                                description="Оберіть в яку систему числення ви хочете переводити число"),
-            view=PersistentViewtest()
-        )
-        await ctx.message.delete()
 
 @Bot.hybrid_command(name = "in2", with_app_command = True, description = "Перевести число в двійкову систему числення")
 @app_commands.guilds(discord.Object(id = "1020640631175004160"))
